requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
requests
loguru
flasgger